% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawProt.R
\name{drawProt}
\alias{drawProt}
\title{Draw lollipop plot for matching variants}
\usage{
drawProt(table, protein_id, organisms)
}
\arguments{
\item{table}{data frame consisting of matching variants, the output of \code{orthoFind} function.}

\item{protein_id}{IDs of proteins to be drawn.}

\item{organisms}{Name of organisms that protein ids belong to.}
}
\value{
plot
}
\description{
This function takes matching variant table, protein ids to be plotted and species names of these protein ids and exports a
lolliplot.
}
