% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoFind.R
\name{orthoFind}
\alias{orthoFind}
\title{This function can be used to find orthologous variants between species. \code{msa} file generated by \code{orthoMSA} function is used as an
input. Each row in output data frame represents a matching variants (orthoVar).}
\usage{
orthoFind(df1, df2, org1, org2, msa, ort = TRUE)
}
\arguments{
\item{df1}{data frame consisting of variant data for organism 1}

\item{df2}{data frame consisting of variant data for organism 2}

\item{org1}{scientific name of the organism organism 1}

\item{org2}{scientific name of the organism organism 2}

\item{msa}{data frame which is the output of \code{orthoMSA} function.}

\item{ort}{Should data be filtered according to type of variant (conserved and non-conserved).
Default is TRUE. Refer to paper for detailed explanation: \url{https://doi.org/10.1101/2021.01.07.424951}}
}
\value{
dataframe object
}
\description{
This function can be used to find orthologous variants between species. \code{msa} file generated by \code{orthoMSA} function is used as an
input. Each row in output data frame represents a matching variants (orthoVar).
}
