% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoMSA.R
\name{orthoMSA}
\alias{orthoMSA}
\title{Create multiple sequence alignment table for use in orthoFind function}
\usage{
orthoMSA(
  species1 = "Homo sapiens",
  species,
  seqFile1 = NA,
  seqFiles = NA,
  customOrt = NaN,
  annot = "ncbi",
  ...
)
}
\arguments{
\item{species1}{Name of the main species name which will be used in msa. Default is \verb{Homo sapiens}.}

\item{species}{A character string or character vector specifying other species names used in msa}

\item{seqFile1}{Path of fasta file consisting of protein sequence of first organism. Default is \code{NA}, which downloads the file from NCBI.}

\item{seqFiles}{A character string or character vector specifying path of fasta files consisting of protein sequences of
species indicated in \code{species} parameter. Default is \code{NA}, which downloads files from NCBI.}

\item{customOrt}{data frame consisting of gene orthology data for given species. Default is \code{NA}, which takes data from AllianceGenome.
If your species is not among the ones indicated in output of \code{listSpecies()} function, then use either custom orthology list or set this
to \code{customOrt = "ensembl"} to get orthology data from biomart.}

\item{annot}{source of annotation. Either "ncbi" or "ensembl" can be used. Default is ncbi.}
}
\value{
dataframe object
}
\description{
This function takes species names and other optional parameters as argument and exports a data frame which is suitable to use as input in
orthoFind function to find orthologous variants between species.
By default, gene orthology data is taken from AllianceGenome website and custom orthology data can be used by specifying \code{customOrt}
parameter. For details, please check Readme file.
}
