% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoMSA.R
\name{orthoMSA}
\alias{orthoMSA}
\title{Create multiple sequence alignment table for use in orthoFind function}
\usage{
orthoMSA(
  species1 = "Homo sapiens",
  species,
  humanSeqFile = NA,
  seqFiles = NA,
  customOrt = NA,
  annot = "ncbi"
)
}
\arguments{
\item{species1}{Name of the main species name which will be used in msa. Default is \verb{Homo sapiens}.}

\item{species}{A character string or character vector specifying other species names used in msa}

\item{humanSeqFile}{Path of fasta file consisting of human protein sequences. Default is \code{NA}, which downloads file from NCBI.}

\item{seqFiles}{A character string or character vector specifying path of fasta files consisting of protein sequences of other
species. Default is \code{NA}, which downloads files from NCBI.}

\item{customOrt}{data frame consisting of gene orthology data for given species. Default is \code{NA}, which takes data from AllianceGenome.}

\item{annot}{source of annotation. Either "ncbi" or "ensembl" can be used. Default is ncbi.}
}
\value{
dataframe object
}
\description{
This function takes species names and other optional parameters as argument and exports a data frame which is suitable to use as input in
orthoFind function to find orthologous variants between species.
By default, gene orthology data is taken from AllianceGenome website and custom orthology data can be used by specifying \code{customOrt}
parameter. For details, please check Readme file.
}
